refactor: 도서관 Writer 설정 클래스 분리
