+++ src/main/java/com/bookspot/batch/library/LibraryStepConfig.java
@@ -9,7 +9,6 @@ import org.springframework.batch.core.step.builder.StepBuilder;
 import org.springframework.batch.core.step.tasklet.MethodInvokingTaskletAdapter;
 import org.springframework.batch.extensions.excel.poi.PoiItemReader;
 import org.springframework.batch.item.database.JdbcBatchItemWriter;
-import org.springframework.batch.item.database.builder.JdbcBatchItemWriterBuilder;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.transaction.PlatformTransactionManager;
@@ -21,17 +20,18 @@ import javax.sql.DataSource;
 public class LibraryStepConfig {
     private final JobRepository jobRepository;
     private final PlatformTransactionManager platformTransactionManager;
-    private final DataSource dataSource;
 
     private final PoiItemReader<Library> libraryExcelReader;
     private final LibraryFileDownloader downloader;
+
+    private final JdbcBatchItemWriter<Library> libraryWriter;
 
     @Bean
     public Step libraryStep() {
         return new StepBuilder(LibraryStepConst.STEP_NAME, jobRepository)
                 .<Library, Library>chunk(LibraryStepConst.LIBRARY_CHUNK_SIZE, platformTransactionManager)
                 .reader(libraryExcelReader)
-                .writer(libraryWriter())
+                .writer(libraryWriter)
                 .build();
     }
 
@@ -63,26 +63,6 @@ public class LibraryStepConfig {
         adapter.setTargetObject(downloader);
         adapter.setTargetMethod("delete");
         return adapter;
-    }
-
-    @Bean
-    public JdbcBatchItemWriter<Library> libraryWriter() {
-        return new JdbcBatchItemWriterBuilder<Library>()
-                .dataSource(dataSource)
-                .sql("""
-                        INSERT INTO library (name, library_code, location, address, updated_at) VALUES
-                        (?, ?, ST_GeomFromText(CONCAT('POINT(', ?, ' ', ?, ')'), 4326), ?, NOW())
-                        ON DUPLICATE KEY UPDATE name = VALUES(name), location = VALUES(location), address = VALUES(address), updated_at = NOW();
-                        """)
-                .itemPreparedStatementSetter(
-                        (library, ps) -> {
-                            ps.setString(1, library.getName());
-                            ps.setString(2, library.getLibraryCode());
-                            ps.setDouble(3, library.getLatitude());
-                            ps.setDouble(4, library.getLongitude());
-                            ps.setString(5, library.getAddress());
-                        })
-                .build();
     }
 
 }
